[IMP] l10n_in_*: Divide Fiscal Localisation for l10n_in
- Introduced two separate fiscal localizations for India:
  - India - Basic: Covers normal accounting features without taxes (no GST,
                   TDS/TCS).
  - India - Advanced: Includes comprehensive tax handling (GST, TDS/TCS).

- Users can switch from India - Basic to India - Advanced. However, once
  switched to India - Advanced, users cannot revert to India - Basic to
  ensure consistency in tax reporting and compliance.

task - 4356554
