+++ addons/l10n_in/data/template/account.account-in_adv.csv
@@ -1,0 +1,28 @@
+"id","name","code","account_type","tag_ids","reconcile","l10n_in_tds_tcs_section_id"
+"p10051","SGST Receivable","10051","asset_current","l10n_in.sgst_tag_account","False",""
+"p10052","CGST Receivable","10052","asset_current","l10n_in.cgst_tag_account","False",""
+"p10053","IGST Receivable","10053","asset_current","l10n_in.igst_tag_account","False",""
+"p100531","IGST Paid on SEZ/Export Sales","100531","asset_current","l10n_in.igst_tag_account","False",""
+"p10054","IGST SEZ/Export Control Account","10054","asset_current","","False",""
+"p10057","Reverse Charge GST on Purchase","10057","asset_current","","False",""
+"p10058","TDS Receivable","10058","asset_current","","False",""
+"p10059","Tax Current Account - Receivable","10059","asset_current","","False",""
+"p11231","TDS Payable","11231","liability_current","","False",""
+"p11232","SGST Payable","11232","liability_current","l10n_in.sgst_tag_account","False",""
+"p11233","CGST Payable","11233","liability_current","l10n_in.cgst_tag_account","False",""
+"p11234","IGST Payable","11234","liability_current","l10n_in.igst_tag_account","False",""
+"p112341","IGST SEZ/Export Control Account","112341","liability_current","","False",""
+"p11237","GST RCM Control Account","11237","liability_current","","False",""
+"p11239","Tax Current Account - Payable","11239","liability_current","","False",""
+"p11244","TDS Deducted","11244","liability_current","","False",""
+"p11245","TCS Collected","11245","liability_current","","False",""
+"p10055","CESS Receivable","10055","asset_current","l10n_in.cess_tag_account","False",""
+"p10056","Tax Receivable","10056","asset_current","","False",""
+"p11235","CESS Payable","11235","liability_current","l10n_in.cess_tag_account","False",""
+"p11236","Tax Payable","11236","liability_current","","False",""
+"p100595",TDS (Withholding Control),"100595","asset_current","","False",""
+"p211210","Professional Fees","211210","expense","","False","l10n_in.tds_section_194j"
+"p211220","Audit Fees","211220","expense","","False","l10n_in.tds_section_194j"
+"p211230","Job Work Expense","211230","expense","","False","l10n_in.tds_section_194c"
+"p211240","Advertisement Expense","211240","expense","","False","l10n_in.tds_section_194c"
+"p211250","Commission/Brokerage Expense","211250","expense","","False","l10n_in.tds_section_194h"
