+++ addons/l10n_in/demo/demo_company.xml
@@ -31,7 +31,7 @@
 
     <function model="account.chart.template" name="try_loading">
         <value eval="[]"/>
-        <value>in</value>
+        <value>in_adv</value>
         <value model="res.company" eval="obj().env.ref('base.demo_company_in')"/>
         <value name="install_demo" eval="True"/>
     </function>
