+++ addons/l10n_in/models/__init__.py
@@ -1,5 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from . import template_in
+from . import template_in_adv
 from . import account_invoice
 from . import account_move_line
 from . import account_payment
