+++ addons/l10n_in/views/res_partner_views.xml
@@ -30,6 +30,9 @@
                             type="object"/>
                 </div>
             </xpath>
+            <xpath expr="//group[@name='fiscal_information']" position="attributes">
+                <attribute name="invisible">not l10n_in_is_advanced_template and country_code == 'IN'</attribute>
+            </xpath>
         </field>
     </record>
